/favorite, /follower
переніс логіку favorite recipes до recipes. Додав перевірки на додавання фолловера
