+++ routes/recipes-router.js
@@ -1,7 +1,6 @@
 import { Router } from "express";
-import recipesControllers from "../controllers/recipesControllers.js";
-import recipesController from "../controllers/favoriteRecipesControllers.js";
 import authenticate from "../middleware/authenticate.js";
+import recipesControllers from "../controllers/recipesControllers.js";
 
 const recipesRouter = Router();
 
@@ -9,7 +8,7 @@ recipesRouter.use(authenticate);
 
 recipesRouter.get("/", recipesControllers.getAllRecipes);
 
-recipesRouter.get("/favorite", recipesController.getFavoriteRecipes);
+recipesRouter.get("/favorite", recipesControllers.getFavoriteRecipes);
 
 recipesRouter.get("/:id", recipesControllers.getOneRecipe);
